tentative addition for/about Cl
Introduction of chlorine.  Tested with an .xyz about chloroethane
(`CCCl`) generated by openbabel, the chlorine atom is depicted as
sphere, separated from the backbone of the molecule.  Because of
a similar observation for compounds with C-S bonds, and because the
analogue CCF was rendered well, perhaps the assignment where to
draw a bond faces an obstacle here.