organize atom definition more like a dictionary
The previous approach requires four lines each for every atom type
as an instance of class Atom.  I think an approach which uses the
element symbol as key to a the corresponding values of mass,
radiosity, and rgb coordinates in a dictionary is - in comparison
to this one - easier to read, and to maintain.  Lacking any better
inspiration, atoms with atom types not yet explicitly defined in
said dictionary default to hydrogen.  In such an instance, the
conversion will not stop for good, however is going to issue one
warning for each atom not fully defined (yet).