test a radii based presence/absence check for bonds
The previous definition of bonds if distances are less than
1.6 A is too simple.  For one, there are covalent bonds which
exceed this threshold (C-I).  For two, simple increase of this
limit to say 1.8 A can cause geminal hydrogens on methylene
units to be assigned a H-H bond in addition to an already
existing C-H bond.

This commit ventures out to add the property dictionary of the
atoms additional values of covalent radii, and the standard
deviations of these covalent radii.  Eventually, there shall be
a comparison between the distance as computed reading the .xyz
file, and a theoretical threshold then used to assign a bond,
or not.  For now, the implementation is after the 1.6 A limit
and only considers C and H as test subjects.  Hence, the test
is not yet as effective as wanted.