provide test molecule with alkyl chain
In previous versions of the script, there was the uniform
threshold of interatomic distances to check if two atoms
could form a bond (any kind of covalent bond).  With 1.6 A,
this was fine for C-C, and C-H, but too short e.g. for
C-S, or C-I.  An increase to 1.8 A rendered the display of
the longer bonds possible; however, simultaneously built
H-H bonds for geminal hydrogens on alkyl groups.

To test the new approach based on comparison of covalence
radii, benzene.xyz now is complemented by pentylbenzene.xyz
to extend the coverage of tests possible.