extend coverage of atomic radii
The coverage of atomic radii is extended -- two periods
H to Ar, plus the more frequently encountered Br and I.
More as a proof of 'a priori, the concept is working',
than backed by chemical evidence for such a compound,
an example with C, H, N, O, P, S, and I is added as a
colorful test model.  Coverage should be enough for many
simpler molecular compounds.

Lines of the chemical dictionary incomplete for covalent
radii are currently muted which doesn't affect the output of
the other elements.