ci: increase default maximal line length to 90
Likely, the array defining the atom properties will never fit
into 79 characters per line recommended by PEP styles.  Hence
the the critical threshold flake8 uses was increased slightly.
There still are a couple of lines exceeding beyond this limit;
at present, it is not obvious if there is a reformat retaining
their function and ease to be read.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>